Support update all paramters as once.
