+++ generator.yaml
@@ -312,15 +312,13 @@ resources:
         from:
           operation: DescribeParameters
           path: Parameters
+    update_operation:
+      custom_method_name: customUpdate
     hooks:
       sdk_read_many_post_set_output:
         template_path: hooks/parametergroup/sdk_read_many_post_set_output.go.tpl
       sdk_create_post_set_output:
         template_path: hooks/parametergroup/sdk_create_post_set_output.go.tpl
-      sdk_update_post_set_output:
-        template_path: hooks/parametergroup/sdk_update_post_set_output.go.tpl
-      sdk_update_pre_build_request:
-        template_path: hooks/parametergroup/sdk_update_pre_build_request.go.tpl
     renames:
       operations:
         CreateParameterGroup:
