--- templates/hooks/parametergroup/sdk_update_post_set_output.go.tpl
@@ -1,5 +1,0 @@
-    ko, err = rm.setParameters(ctx, ko)
-
-    if err != nil {
-        return nil, err
-    }
